# service-account/docs/openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8661",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "계정 API",
      "description": "계정 관리를 위한 API"
    }
  ],
  "paths": {
    "/{id}": {
      "get": {
        "tags": [
          "계정 API"
        ],
        "summary": "ID로 계정 세부 정보 조회",
        "description": "ID를 사용하여 특정 계정의 세부 정보를 조회합니다.",
        "operationId": "getAccountById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "조회할 계정의 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "200": {
            "description": "계정 세부 정보 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReadAccountResponse"
                }
              }
            }
          },
          "404": {
            "description": "계정을 찾을 수 없습니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReadAccountResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReadAccountResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "계정 API"
        ],
        "summary": "기존 계정 업데이트",
        "description": "기존 계정을 업데이트합니다.",
        "operationId": "updateAccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "업데이트할 계정 정보",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "잘못된 요청"
          },
          "200": {
            "description": "계정이 성공적으로 업데이트되었습니다."
          },
          "404": {
            "description": "계정을 찾을 수 없습니다."
          },
          "500": {
            "description": "서버 오류"
          }
        }
      },
      "delete": {
        "tags": [
          "계정 API"
        ],
        "summary": "계정 삭제",
        "description": "기존 계정을 삭제합니다.",
        "operationId": "deleteAccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "삭제할 계정의 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "204": {
            "description": "계정이 성공적으로 삭제되었습니다."
          },
          "404": {
            "description": "계정을 찾을 수 없습니다."
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "계정 API"
        ],
        "summary": "모든 계정 목록 조회",
        "description": "모든 계정의 목록을 페이지네이션하여 조회합니다.",
        "operationId": "getAccounts",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "description": "페이지네이션 정보",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "200": {
            "description": "계정 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReadAccountsResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageReadAccountsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "계정 API"
        ],
        "summary": "새 계정 생성",
        "description": "새로운 계정을 생성합니다.",
        "operationId": "createAccount",
        "requestBody": {
          "description": "생성할 계정 정보",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "잘못된 요청"
          },
          "201": {
            "description": "계정이 성공적으로 생성되었습니다.",
            "headers": {
              "Location": {
                "description": "생성된 계정의 URI",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          }
        }
      },
      "UpdateAccountRequest": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "CreateAccountRequest": {
        "required": [
          "email",
          "name"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "ReadAccountResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ReadAccountsResponse": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PageReadAccountsResponse": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReadAccountsResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "paged": {
            "type": "boolean"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "direction": {
            "type": "string"
          },
          "nullHandling": {
            "type": "string"
          },
          "ascending": {
            "type": "boolean"
          },
          "property": {
            "type": "string"
          },
          "ignoreCase": {
            "type": "boolean"
          }
        }
      }
    }
  }
}